Removed the remove operation
Turns out, destructive operations are ridiculous.  I even noted the
problem in the comments for the method!  It 'skips' the next token. Like
that wasn't about four thousand bugs waiting to happen.

So anyway, it now outputs the tokens to a new list.  I should probably
remove the token list global thingy, but whatevs.  More stuff to free.

Also, everything is broken.