Next step towards replacing macro values in #if
Made the mcc_CopyToken function complete - apparently uninitialised
values make a mess of things when you try to use them.

I also shifted the macro replacement functions around a little such that
they return a copy of the resultant tokens.  I will likely need to make
this a little more general still, but this is a good first step.