A couple more compiler defined macros
This gets me far enough to know that I'm handling macro functions all
wrong.  The arguments list is an expression, complete with as many sets
of parentheses as your heart desires.  In fact, adding more is probably
safe, given that we don't know what kind of expression is going to be
passed in.  Sigh.