Parse macro function arguments as integer expressions
Function calls can have expressions as their arguments, e.g.
ADD(2*4, 2). So we really need to parse all function arguments as though
they are expressions, even if that expression is just a single number.

I have accomplished that, but the memory handling is all borked, which
needs to be fixed before the thing will run.