Started supporting commas in expressions
Don't yet know how it helps me, other than the face that I don't throw
an error saying that a comma is unrecognised as part of a function
argument.  Unfortunately, macro functions are weird, and what looks like
two arguments to the reader, is actually a single argument, and can't be
evaluated - should be a textual replacement, only.  Ugh.