Make replace current data take a list
The use case has gotten a little weird, so I should probably clean that
up at some point, but as a dirty trick to move forward, this allows me
to insert a list of tokens at the 'current point' which is super handy.
I also get to replace the 'current' token, which is also handy.  It's
not something I could explain generally, though.