Hard-coded the version specific gcc-deployed header dir
Compilers apparently include a bunch more stdlib stuff separate from the
base glibc than I thought.  Either that, or the stdlib is split into
callable functions and things that need to be compiled in, like macros,
atomics, etc and the compiler provides the latter.  Or something.

Either way, I now have a hard dependency on gcc being installed and in a
certain dir, at a certain version (8) and I don't have a good solution
for that.