Make getting a non-whitespace token first class
It seems we need it in a lot of places - in fact, the main reason I can
think of where we don't need to ignore whitespace is when we need to
differentiate between a macro and a function-macro in a #define.

I also updated a bunch of the doxygen comments, because when I
originally wrote them, I didn't realise that that's what they were and
didn't have them well-formed.  Stupid non-reviewing practices.