Add parser admin functions
It is a bit of a clone-and-own from the ICE parser, and I should see if
I want to de-duplicate those two implementations.  Especially as they
export competing definitions of mcc_AST_t and mcc_ASTNode_t which
haven't collided yet, because the actual parser isn't hooked up to the
main program.