Move macro expansion to their own files
Turns out, the idea that the macro expansion code is so heavy-weight was
all tied up in how much they were placed with the main preprocessor
functions.  Once we move them to their own spot, we can see it's only a
couple of (still relatively large) functions to deal with, but more
manageable chunks, for sure.