Theoretically refactor function replace to return tokens
Before we refactor function macro expansion to work on a generic token
list, it would be helpful to return the changes as a new list.  This
means we can avoid the in-place wrangling that hurts my head and will
make it possible to call the function recursively and have it "just
work", right?