Fix basic macro function expansion
Refactored the expansion to use proper lists, creating a new one for
each iteration, which is a little less efficient, especially in terms of
many small memory allocations, but the code is a lot easier to reason
about and behaves a bit more like you might expect.

So theoretically, I just need to extract the macro function parsing code
and I'll be able to handle macro functions in functions... getting me a
long way towards correct preprocessing?